+++ lib/default_fields/name_field.dart
@@ -1,3 +1,4 @@
+import 'package:championforms/models/field_colors.dart';
 import 'package:championforms/models/field_types/compound_field.dart';
 import 'package:championforms/models/field_types/formfieldclass.dart';
 import 'package:championforms/models/field_types/textfield.dart';
@@ -94,6 +95,8 @@ class NameField extends CompoundField {
     super.hideField,
     super.rollUpErrors,
     super.theme,
+    super.colors,
+    super.animateValidationErrors,
     super.validators,
     super.validateLive,
     super.onSubmit,
@@ -110,6 +113,8 @@ class NameField extends CompoundField {
     bool? hideField,
     bool? rollUpErrors,
     FormTheme? theme,
+    FieldColors? colors,
+    bool? animateValidationErrors,
     List<Validator>? validators,
     bool? validateLive,
     Function(FormResults results)? onSubmit,
@@ -140,6 +145,9 @@ class NameField extends CompoundField {
       hideField: hideField ?? this.hideField,
       rollUpErrors: rollUpErrors ?? this.rollUpErrors,
       theme: theme ?? this.theme,
+      colors: colors ?? this.colors,
+      animateValidationErrors:
+          animateValidationErrors ?? this.animateValidationErrors,
       validators: validators ?? this.validators,
       validateLive: validateLive ?? this.validateLive,
       onSubmit: onSubmit ?? this.onSubmit,