+++ lib/models/field_types/compound_field.dart
@@ -103,6 +103,8 @@ abstract class CompoundField extends Field with TextFieldConverters {
     super.hideField,
     this.rollUpErrors = false,
     super.theme,
+    super.colors,
+    super.animateValidationErrors,
     super.validators,
     super.validateLive,
     super.onSubmit,