+++ lib/models/field_types/convienence_classes/chipselect.dart
@@ -1,4 +1,5 @@
 import 'package:championforms/models/field_builder_context.dart';
+import 'package:championforms/models/field_colors.dart';
 import 'package:championforms/models/field_types/optionselect.dart';
 import 'package:championforms/models/multiselect_option.dart';
 import 'package:championforms/models/themes.dart';
@@ -32,6 +33,8 @@ class ChipSelect extends OptionSelect {
     super.onChange,
     super.fieldLayout,
     super.fieldBackground,
+    super.colors,
+    super.animateValidationErrors,
     Widget Function(FieldBuilderContext)? fieldBuilder,
   }) : super(fieldBuilder: fieldBuilder ?? checkboxFieldBuilder);
 
@@ -44,6 +47,8 @@ class ChipSelect extends OptionSelect {
     Widget? leading,
     Widget? trailing,
     FormTheme? theme,
+    FieldColors? colors,
+    bool? animateValidationErrors,
     String? title,
     String? description,
     bool? disabled,
@@ -79,6 +84,9 @@ class ChipSelect extends OptionSelect {
       leading: leading ?? this.leading,
       trailing: trailing ?? this.trailing,
       theme: theme ?? this.theme,
+      colors: colors ?? this.colors,
+      animateValidationErrors:
+          animateValidationErrors ?? this.animateValidationErrors,
       title: title ?? this.title,
       description: description ?? this.description,
       disabled: disabled ?? this.disabled,